Assign_and_Route.py is a script that replaces assignriverelev and the two routing scripts: routestreamnetwork and route river cells.  Speed is much improved because I eliminated the queries to the NHD.
Finalize_SFR_revised_OCT.py is a new 'final' script.  The approach is to keep
all segments so there are multiple segments in a cell.  This keeps the
routing clean.  If there are multiple segments, the one with the greatest
conductance is identified as the dominant segment and its properties are
used in the SFR output.  For the others, the Kstreambed is set to a very
small value.  The remaining problem is that there may be multiple
segments in a cell that are actually routed and could be merged
together to get a better estimate of the stream length in the cell.
The routing seems much better in this version though.