Added RemoveSegmentGaps.ipynb -- an iPython notebook for dealing with the ANNOYING fact that when you manually remove SFR segments, you need to renumber them all so that there are no gaps in the number sequence.
To run, navigate to the directory and, at the command line type "ipython notebook" and a browser should open with the notebook in it.

You then need to run each cell in order from the top (hit "shift + enter" while in the cell) and it will either run and advance to the next cell or will throw an exception.